Update mongoose schema for German verbs and set up a trigger to populate the database. (#29)
* feat: start updating mongoose schema for german and english

Set up a test to start pumping a simple german verb through a transformer.

* add model for german

* feat: add particle to mongo schema

Added particle to Mongo schema in prep for moving all seperable verbs into variations of the primary
verb.

* feat: wip fix mongo push to db

Successfully pushed to the DB. The schema is not populating correctly.

* feat: fix schema with help from openai

Fixed schema and tested it against the db.

* test: add a simple test for schema conversion

Added a simple test for ab|blenden to guard against accidental schema changes.

* feat: move insert functionality to trigger on a button click

Added a button to the view to trigger a put for German. Commented out the insert in Populte Routes.